+++ src/components/CardGrid/views.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const CardGridFlexContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+  padding: 0px;
+  width: 100%;
+  max-width: 1224px;
+  margin-top: 48px;
+  margin-left: 36px;
+`;